backend/commit_getter_service/requirements.txt
Flask
boto3
PyGithub
moto
